requirements.txt
Django~=4.0
djangorestframework~=3.13.0
Pillow~=8.4.0
django-cors-headers~=3.10.1